desafio_dos_numeros_backend/requirements.txt
Django==5.2.1
djangorestframework==3.15.1
celery==5.3.6
psycopg2-binary